extract inMemoryDataSource into ocrcommon
